requirements.txt
numpy==1.18.1
pandas~=1.0.3
scikit-learn~=0.22.1
uvicorn~=0.13.4
joblib~=0.15.1
python-dotenv~=0.15.0
typing~=3.7.4.3
setuptools~=47.1.1
fastapi~=0.63.0
starlette==0.13.6
category_encoders
gunicorn